Add a sync_open_prs management command.
Sometimes the PRs the git server and the PRs CIVET
thinks are open get out of sync due to failed events.
This automatically closes PRs that CIVET thinks
are open but have been closed on the Git server.
Also reports any open PRs on the Git server
that CIVET doesn't have open.
Added as a management command as this is mainly for admins.

closes #331