Easier to use Git API.
This allows for creating GitAPI objects with either
an existing user or just with a token.
Right now we don't have any live testing of the APIs,
all the "get", "post", etc, calls are mocked.
This makes testing/debugging the various API calls easier to
setup and use from a python shell.