Fix invalidate event.
When invalidating all jobs on an event, it didn't take
into account any job dependencies, it just started all the jobs.
Had the case where an event was invalidated and the "Precheck"
failed but all the other jobs were running. Updating the status
on GitHub got all weird because the checks we were doing said
that none of those jobs should be running.