Civet, or perhaps GitHub is confused
Civet or Github is demanding a set of tests WHICH NO LONGER EXIST
to be completed before I can merge. The civet recipes in question
have been deleted because they are python2-only tests.

So, instead of branching off my previous work (which had python2
tests available on civet at the time), start anew (in a very dumb
way of copy/paste contents out of repo!), and try again.